launchkit_api/requirements.txt
asgiref==3.8.1
Django==5.1.7
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
pillow==11.2.1
PyJWT==2.9.0
sqlparse==0.5.3
tzdata==2025.2
